Move LoadWeakLibEosTableFull tests to test_weaklib_eos_loader.cpp
Relocate WeakLib EOS full-table tests from the generic HDF5 loader test
file to the dedicated WeakLib EOS loader test file where related tests
already exist. This improves test organization by grouping all EOS
loader tests together.

Changes:
- Move LoadWeakLibEosTableFull and MakeDeviceCopy tests
- Move helper functions: WriteIntArrayDataset, WriteDoubleArrayDataset,
  WriteStringArrayDataset, WriteDoubleArray3dDataset, WriteIntArray3dDataset
- Add CreateWeakLibEosTestFileFull (renamed from CreateWeakLibEosTestFile)
- Remove unused includes from test_hdf5_loader.cpp
